requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
pytz==2023.3